+++ src/ChronicleServiceProvider.php
@@ -19,6 +19,7 @@ use Chronicle\Console\Commands\ReportCommand;
 use Chronicle\Console\Commands\ShowEntryCommand;
 use Chronicle\Console\Commands\StatsCommand;
 use Chronicle\Console\Commands\SubjectEraseCommand;
+use Chronicle\Console\Commands\SubjectKeysCommand;
 use Chronicle\Console\Commands\VerifyEntryCommand;
 use Chronicle\Console\Commands\VerifyExportCommand;
 use Chronicle\Context\QueueJobContext;
@@ -115,6 +116,7 @@ class ChronicleServiceProvider extends ServiceProvider
                 AnchorVerifyCommand::class,
                 LegalHoldCommand::class,
                 SubjectEraseCommand::class,
+                SubjectKeysCommand::class,
             ]);
         }
     }